requirements.txt
# Chicago Event Monitor - Minimal Dependencies
# Only what we actually need for Phase 1

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Email notifications
sendgrid>=6.11.0

# Environment variables
python-dotenv>=1.0.0

# Note: Removed from original requirements.txt:
# - selenium (no JavaScript rendering needed)
# - schedule (using cron instead)
# - pytest, black, flake8 (add later if needed)
